+++ recvs/base_test.go
@@ -1,21 +1,21 @@
 package recvs_test
 
 import (
-	"net/http"
 	"sync"
-	"time"
+
+	"github.com/Laisky/zap"
 
 	"github.com/Laisky/go-fluentd/libs"
 	utils "github.com/Laisky/go-utils"
 )
 
 var (
-	httpClient = &http.Client{ // default http client
-		Transport: &http.Transport{
-			MaxIdleConnsPerHost: 20,
-		},
-		Timeout: 30 * time.Second,
-	}
+	// httpClient = &http.Client{ // default http client
+	// 	Transport: &http.Transport{
+	// 		MaxIdleConnsPerHost: 20,
+	// 	},
+	// 	Timeout: 30 * time.Second,
+	// }
 	counter = utils.NewCounter()
 	msgPool = &sync.Pool{
 		New: func() interface{} {
@@ -28,5 +28,7 @@ var (
 )
 
 func init() {
-	utils.SetupLogger("debug")
+	if err := utils.Logger.ChangeLevel("debug"); err != nil {
+		utils.Logger.Panic("change log level", zap.Error(err))
+	}
 }