+++ senders/elasticsearch_test.go
@@ -1,15 +1,10 @@
 package senders_test
 
-import (
-	"net/http"
-	"time"
-)
-
 var (
-	httpClient = &http.Client{ // default http client
-		Transport: &http.Transport{
-			MaxIdleConnsPerHost: 20,
-		},
-		Timeout: 30 * time.Second,
-	}
+// httpClient = &http.Client{ // default http client
+// 	Transport: &http.Transport{
+// 		MaxIdleConnsPerHost: 20,
+// 	},
+// 	Timeout: 30 * time.Second,
+// }
 )